requirements.txt
openai-agents[voice]>=0.6.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.24.0
